+++ CHANGELOG.md
@@ -5,6 +5,9 @@ corresponding version of the Dune core modules.
 
 ## Master (will become release 2.10)
 
+- The class `SizeInfo` and the corresponding header `functionspacebases/sizeinfo.hh`
+  have been deprecated. One now can directly use `vectorBackend.resize(globalBasis)`
+  instead of `vectorBackend.resize(sizeInfo(globalBasis))`.
 - The deprecated header `common/treedata.hh` was removed.
 - The deprecated header `common/referencehelper.hh` was removed. Use the corresponding header
   from dune-common instead.